refactor: comment Future feat with restore User session
